make confusing abstractions
